move number parsers to dedicated module
