+++ crates/hcl-edit/src/parser/expr.rs
@@ -1,7 +1,8 @@
+use super::number::number;
 use super::{
     cut_char, cut_ident, cut_tag, decor,
     error::{Context, Expected, InternalError},
-    ident, line_comment, number, prefix_decor, raw, sp, spanned, str_ident, string, suffix_decor,
+    ident, line_comment, prefix_decor, raw, sp, spanned, str_ident, string, suffix_decor,
     template::{heredoc_template, string_template},
     ws, IResult, Input,
 };
